TrainAndTest/requirements.txt
torch>=1.6
onnxruntime>=1.7.0
numpy
pandas
fair-esm
pytorch-lightning
scikit-learn